It frees it's own memory now (note ...
that having (&data_ptr) was just me changing random things to see
where the value reported by valgrind changed if I did random things
to the parameters.

It should have been (data_ptr), but the cool thing about leaving
that was that when I did set the NPY_OWNDATA flag, the destruction
of the numpy array caused a segfault.  Proving that it was freeing
something.  Setting the correct parameter (data_ptr) fixed the crash
pretty much proving that a call to free(data_ptr) is happening.

Running the memory test of creating and destroying such arrays
repeatedly and looking at the memory usage of the process shows that
there is no memory leakage going on.